refactor(search): split search.go into fusion/ranking/enrich/fallback; decompose Search
Split the 1013-line search.go along its concern clusters: the engine and
the Search orchestration stay in search.go; RRF fusion, the score-mutating
ranking passes, graph enrichment, and the substring fallback move to
fusion.go, ranking.go, enrich.go, fallback.go. All within one package, so
the moves are behavior-preserving.

Decompose Search (gocyclo 37) into a named pipeline over a per-invocation
searchContext: prepareSearch -> fuseQueries (with retrieveCandidates) ->
rankResults -> enrichResults -> finalizeResults, each at gocyclo <=15. The
ranking-pass call ORDER is preserved exactly (each pass mutates scores the
next reads). Also decompose promoteParents (gocyclo 18) into groupByParent
and selectPromotions, retiring its ledger entry.

No weight, threshold, or call-order change. score-guard/shape/normalize
tests green; maket search query byte-identical on the existing index;
package coverage 95.3% -> 95.8%.
